requirements.txt
# --- Phase 1: Data Collection ---
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
pandas>=2.2.0

# --- Phase 2: Storage ---
psycopg2-binary>=2.9.9
pymongo>=4.6.0
